+++ signal-query-api/src/services/signals.service.ts
